refactored to account for undefined options
