+++ src/utils/getCommands.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const getAllFiles = require('./getAllFiles');
 
-module.exports = (exceptions = []) =>{
+const getLocalCommands = (exceptions = []) =>{
     let localCommands = [];
 
     const commandCategories = getAllFiles(
@@ -23,4 +23,27 @@ module.exports = (exceptions = []) =>{
         }
     }
     return localCommands;
+};
+
+
+//gets the application commands
+const getApplicationCommands = async(client, guildId)=>{
+    let applicationCommands;
+
+    //if guild
+    if(guildId){
+        const guild = await client.guilds.fetch(guildId);
+        applicationCommands = guild.commands;
+    }
+    else{
+        applicationCommands = await client.application.commands;
+    }
+    await applicationCommands.fetch();
+    
+    return applicationCommands;
+};
+
+module.exports = {
+    getApplicationCommands,
+    getLocalCommands,
 };
